requirements.txt
pydantic~=1.8.2
pymongo~=3.12.1
motor~=2.5.1
fastapi~=0.70.0
passlib~=1.7.4
uvicorn~=0.15.0
starlette~=0.16.0
python-jose~=3.3.0
modzy-sdk~=0.5.7
numpy~=1.21.3
opencv-python~=4.5.4.58